joining, leaving, kicked out of team
All of these are now working. they have conditionals such that if you are already a part of a team you can't join another one. you can't leave a team if you are the admin for that team. admins can remove people from their teams.

I altered the teams data structure to make checking if a team is full and who the admin of the team is a lot easier..