All routes complete
I have basically finished everything that i can think of aside from the optional routes. Things seems to be working fairly well with few bugs. If anything, we should probably implement an error pug page since we reference it a lot within the code when our mongo database calls arise in error, but if something were to happen, the app would crash because this page doesn't exist.